Refactor K8s tests - move RenderingContext
[#171343212]
