+++ k8s/matchers/produce_yaml_matcher.go
@@ -11,20 +11,6 @@ import (
 	"os/exec"
 )
 
-type RenderingContext struct {
-	templates []string
-	data      map[string]string
-}
-
-func (r RenderingContext) WithData(data map[string]string) RenderingContext {
-	r.data = data
-	return r
-}
-
-func NewRenderingContext(templates ...string) RenderingContext {
-	return RenderingContext{templates, nil}
-}
-
 type ProduceYAMLMatcher struct {
 	matcher  types.GomegaMatcher
 	rendered string
